contacts_project_web/requirements.txt
Flask>=1.1.1
google-cloud-ndb
